[TYP] int32 is not always the default

+++ hypothesis-python/src/hypothesis/extra/numpy.py
@@ -904,7 +904,7 @@ def integer_array_indices(
     shape: Shape,
     *,
     result_shape: st.SearchStrategy[Shape] = array_shapes(),
-    dtype: D = np.int32,
+    dtype: D = np.int_,
 ) -> st.SearchStrategy[Tuple[NDArray[D], ...]]:
     """Return a search strategy for tuples of integer-arrays that, when used
     to index into an array of shape ``shape``, given an array whose shape
